Cleanups
I consider them cleanups anyway. ;-)

We don't really need a buffer pool AFAICT. We're not doing anything high
performance. And even if we were I'd rather be doing it with a
sync.Pool.

Use an anonymous struct instead of an ugly (to me anyway)
map[string]interface{} for the template.

Consolidate some vars.